refactor: 기물 공통 로직을 상위 클래스로 올리고 메서드명 개선
- Cannon, Chariot의 중복된 validateRule, resolvePath를
  DiagonalPalaceMovementPiece로 이동
- Soldier의 validateRule, resolvePath를
  ForwardMovementPiece로 이동, validateNormalRule만 하위에서 구현
- validateNotSameTeam → validateTarget으로 메서드명 수정
  (도착 위치 기물 검증이라는 의도를 더 명확히 표현)
