+++ src/main/java/domain/piece/Cannon.java
@@ -1,6 +1,5 @@
 package domain.piece;
 
-import domain.coordination.Coordination;
 import domain.piece.error.PieceException;
 
 import java.util.List;
@@ -23,45 +22,13 @@ public class Cannon extends DiagonalPalaceMovementPiece {
     }
 
     @Override
-    public void validateRule(Coordination from, Coordination to) {
-        if (palaceMovement.isPalace(from, to)) {
-            palaceMovement.validateRule(from, to);
-            return;
-        }
-        validateNormalRule(from, to);
-    }
-
-    private void validateNormalRule(Coordination from, Coordination to) {
-        boolean movable = from.isHorizontal(to) || from.isVertical(to);
-        if (!movable) {
-            throw new PieceException(IMPOSSIBLE_MOVE_MESSAGE);
-        }
-    }
-
-    @Override
-    public List<Coordination> resolvePath(Coordination from, Coordination to) {
-        if (palaceMovement.isPalace(from, to)) {
-            if (from.isDiagonal(to)) {
-                return from.diagonalPathTo(to);
-            }
-        }
-        if (from.isVertical(to)) {
-            return from.verticalPathTo(to);
-        }
-        if (from.isHorizontal(to)) {
-            return from.horizontalPathTo(to);
-        }
-        throw new IllegalStateException(UNRESOLVABLE_PATH_MESSAGE);
-    }
-
-    @Override
     public void validatePath(List<Piece> piecesOnPath) {
         validateExactlyOneBridge(piecesOnPath);
         validateBridgeIsNotCannon(piecesOnPath);
     }
 
     @Override
-    public void validateNotSameTeam(Piece target) {
+    public void validateTarget(Piece target) {
         if (target.isEmpty()) {
             return;
         }
