+++ src/main/java/domain/piece/DiagonalPalaceMovementPiece.java
@@ -3,6 +3,9 @@ package domain.piece;
 import domain.coordination.Coordination;
 import domain.movement.DiagonalPalaceMovement;
 import domain.movement.PalaceMovement;
+import domain.piece.error.PieceException;
+
+import java.util.List;
 
 public abstract class DiagonalPalaceMovementPiece extends Piece {
 
@@ -17,6 +20,33 @@ public abstract class DiagonalPalaceMovementPiece extends Piece {
 
     @Override
     public void validateRule(Coordination from, Coordination to) {
-        palaceMovement.validateRule(from, to);
+        if (palaceMovement.isPalace(from, to)) {
+            palaceMovement.validateRule(from, to);
+            return;
+        }
+        validateNormalRule(from, to);
+    }
+
+    @Override
+    public List<Coordination> resolvePath(Coordination from, Coordination to) {
+        if (palaceMovement.isPalace(from, to)) {
+            if (from.isDiagonal(to)) {
+                return from.diagonalPathTo(to);
+            }
+        }
+        if (from.isVertical(to)) {
+            return from.verticalPathTo(to);
+        }
+        if (from.isHorizontal(to)) {
+            return from.horizontalPathTo(to);
+        }
+        throw new IllegalStateException(UNRESOLVABLE_PATH_MESSAGE);
+    }
+
+    private void validateNormalRule(Coordination from, Coordination to) {
+        boolean movable = from.isHorizontal(to) || from.isVertical(to);
+        if (!movable) {
+            throw new PieceException(IMPOSSIBLE_MOVE_MESSAGE);
+        }
     }
 }
